moves ntp check to network checks
