+++ pkg/collector/corechecks/network/common_test.go
@@ -1,10 +1,9 @@
-package other
+package network
 
 import (
-	"testing"
+	"github.com/stretchr/testify/mock"
 
 	"github.com/DataDog/datadog-agent/pkg/aggregator"
-	"github.com/stretchr/testify/mock"
 )
 
 //MockSender allows mocking of the checks sender
@@ -50,35 +49,4 @@ func (m *MockSender) Event(e aggregator.Event) {
 //Commit enables the commit mock call.
 func (m *MockSender) Commit() {
 	m.Called()
-}
-
-var ntpCfgString = `
-offset_threshold: 60
-port: ntp
-version: 3
-timeout: 5
-`
-
-var ntpCfg = []byte(ntpCfgString)
-var ntpInitCfg = []byte("")
-
-func TestNTP(t *testing.T) {
-	ntpCheck := new(NTPCheck)
-	ntpCheck.Configure(ntpCfg, ntpInitCfg)
-
-	mockSender := new(MockSender)
-	ntpCheck.sender = mockSender
-
-	mockSender.On("Gauge", "ntp.offset", mock.AnythingOfType("float64"), "", []string(nil)).Return().Times(1)
-	mockSender.On("ServiceCheck",
-		"ntp.in_sync", mock.AnythingOfType("aggregator.ServiceCheckStatus"), "",
-		[]string(nil), mock.AnythingOfType("string")).Return().Times(1)
-
-	mockSender.On("Commit").Return().Times(1)
-	ntpCheck.Run()
-
-	mockSender.AssertExpectations(t)
-	mockSender.AssertNumberOfCalls(t, "Gauge", 1)
-	mockSender.AssertNumberOfCalls(t, "ServiceCheck", 1)
-	mockSender.AssertNumberOfCalls(t, "Commit", 1)
 }
